Update Calendar.html
decided to follow a different tutorial because the original one was 5 hours long. I will go through the 5 hour one but most likely in a seperate project and then once I understand the details of it implement it to another project. The video Im referencing for this calendar is https://www.youtube.com/watch?v=0AHWULS01Pg&t=16s