Update Calendar.html
Added the save event and delete event buttons. They don't have function for now. Will need to find a way to have the weight check be sent to the calendar so there can be a daily weigh in. Also will make a graph tab in the future to track progress over time